Rename state_management folder as view_models
