+++ packages/stacked/lib/src/view_models/base_view_models.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/foundation.dart';
 import 'package:stacked/src/mixins/listenable_service_mixin.dart';
-import 'package:stacked/src/state_management/helpers/data_state_helper.dart';
+import 'package:stacked/src/view_models/helpers/data_state_helper.dart';
 
 import 'helpers/builders_helpers.dart';
 import 'helpers/busy_error_state_helper.dart';
