+++ packages/stacked/lib/src/view_models/data_models/multiple_data_models.dart
