+++ packages/stacked/lib/src/view_models/helpers/form_state_helper.dart
