+++ packages/stacked/lib/src/view_models/helpers/index_tracking_state_helper.dart
