+++ packages/stacked/lib/src/view_models/helpers/message_state_helper.dart
