+++ packages/stacked/lib/src/view_models/selector_view_model_builder_widget.dart
