+++ packages/stacked/lib/src/view_models/ui/skeleton_loader.dart
