+++ packages/stacked/lib/src/view_models/view_model_builder_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
-import 'package:stacked/src/state_management/view_model_builder.dart';
+import 'package:stacked/src/view_models/view_model_builder.dart';
 
 /// A widget that wraps the [ViewModelBuilder] class in a less boiler plate use of the widget
 ///
