+++ packages/stacked/lib/src/view_models/view_model_widget.dart
